+++ CHANGELOG.md
@@ -8,5 +8,6 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-### Changed
+- feature/1200366404543322 - Modification which Integrates validator reset capability / Max depth directory validator
 
+### Changed